Refactor MissionReader to LocalPlanner
