+++ src/isar/apis/missions/list_predefined_missions.py
@@ -3,9 +3,8 @@ from http import HTTPStatus
 
 from flask_restx import Namespace, Resource
 from injector import inject
-
 from isar.config import config
-from isar.services.readers.mission_reader import LocalPlanner
+from isar.mission_planner.local_planner import LocalPlanner
 
 api = Namespace(
     config.get("api_namespaces", "eqrobot_missions_namespace"),
