+++ tests/conftest.py
@@ -1,9 +1,9 @@
 import pytest
 from injector import Injector
-
 from isar import MissionPlannerModule, create_app
 from isar.config.keyvault.keyvault_service import Keyvault
 from isar.mission_planner.echo_planner import EchoPlanner
+from isar.mission_planner.local_planner import LocalPlanner
 from isar.models.communication.queues.queues import Queues
 from isar.modules import (
     CoordinateModule,
@@ -17,7 +17,6 @@ from isar.modules import (
 )
 from isar.services.coordinates.transformation import Transformation
 from isar.services.readers.map_reader import MapConfigReader
-from isar.services.readers.mission_reader import LocalPlanner
 from isar.services.service_connections.request_handler import RequestHandler
 from isar.services.service_connections.stid.stid_service import StidService
 from isar.services.utilities.path_service import PathService
@@ -25,6 +24,7 @@ from isar.services.utilities.scheduling_utilities import SchedulingUtilities
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states import Collect, Idle, Monitor, Send
 from isar.storage.storage_service import StorageService
+
 from tests.test_modules import MockStorageModule
 from tests.test_utilities.mock_interface.mock_robot_interface import MockRobot
